requirements.txt
annotated-types==0.7.0
anyio==4.5.2
cffi==1.17.1
click==8.1.8
cryptography==45.0.4
dnspython==2.7.0
exceptiongroup==1.3.0
fastapi==0.115.12
greenlet==3.1.1
h11==0.16.0
idna==3.10
motor==3.7.1
neo4j==5.28.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pymongo==4.13.1
PyMySQL==1.1.1
pytz==2025.2
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.44.0
typing_extensions==4.13.2
uvicorn==0.33.0
python-dateutil~=2.9.0.post0